Bug fix on branch
OFJ-541 The filter by finding source does not work on summary page

Apparently search and the controller were using sourceId and source paramaters alternatively, and not complying with one. Changed to just be sourceId.
Also changed the query from a leftJoin to an innerJoin on the Source table, which fixes the counts.