Improvement on the trunk
OFJ-358 Implement multiple roles per account                                                                                                                                                                  
                                                                                                                                                                                                               
Implementation mostly complete. Needs further testing and a bit of cleanup (an object or two is being assigned to a view), but seems to be mostly working.                                                    
                                                                                                                                                                                                               
Allows assignment of multiple roles to users, and assignment of multiple organizations to each role assigned to the user.                                                                                     
                                                                                                                                                                                                               
New method getOrganizationsByRole($roleId) added to User model                                                                                                                                                
ACL modified to work with the new models 
Sample data updated to work with new models

Fisma_Form_Element_CheckboxTree modified to support grouping
Fisma_Yui_Form_Button modified to use the value as the value, instead of the label, if it's specified
Modify Fisma_Yui_TabView to support assignment of IDs to tabs
CSS updated for tab view