New feature on trunk OFJ-562 Add an "attach artifacts" behavior
Clean up error handling for attach artifacts behavior, especially if file is blank or over the size limit. Also work around an xdebug bug which will corrupt json responses when the file upload is over the size limit. Lastly, add a close box to the modal dialog which will cancel the in-progress upload.