New feature on trunk OFJ-594 Add modules administration page
Add module admin page. I added a new button style called SwitchButton that basically mimics the iphone's ON/OFF buttons. This is based on a Jquery implementation which is GPL.

This commit provides the user interface and client-side code for modules. The next commit will contain the logic to conditionally enable/disable features based on the module status.