New feature on trunk
OFJ-373 Add "OIG Freeze" feature to incident response module

Change all references to User::currentUser() to $this->_me
Refactor handling of related models in IR, such that the related records are created on Incident->save(), so that the listener on the Incident model fires
Modify AttachArtifacts behavior to call save() on the instance, rather than on the artifact, so that the Lockable listener can pick this up
Add reference checking on the invoker in the Lockable behavior. This is a little bit weird, since Doctrine isn't adhering to it's own API in terms of how references are stored in Doctrine_Record::_references.
Fix the privilege check in the Lockable behavior, to check for 'lock' rather than 'incident_lock'.
Add exception if an unprivileged user tries to modify a record that is related to a currently locked record.