Bug on trunk OFJ-525 Incident model does not contain reporter's IP address
Add reporter IP to model and make sure it is set in the constructor.

Also, resolve the confusion about source IP which actually didn't have a form in the incident report until now.

Finally, there was a weird thing in ACL related to the Lockable behavior which was preventing anonymous user reporting. I modified _isAllowed() to return false if the user is not logged in.