Bug on trunk OFJ-612 Auto complete search box should not be moved downward as type keyword
Not sure what the root cause is. The box moves down because YUI autocomplete has a "snap to container" feature where it repositions the popup automatically. Not sure what the point of that is, so I disabled it. This fixes the moving behavior, but now the menu pops up a little too high. I added some CSS to move it down 15px so that it looks right (at least in the browsers I tested).