Bug on release branch OFJ-640 FISMA report (quarterly and annual) crash Excel 2007
Fix the counting of C&A versus non C&A systems. Every system should fall into one of these two categories, but before we were excluding systems with an expired C&A. This patch will include expired C&As into the non-C&A category, which is correct according to current requirements.