Bug on release branch OFJ-646 The overdue data does not appear on overdue report list
This appears to be caused by using the wrong constructor, i.e. DataSource() when LocalDataSource() is actually correct.

The code is so complicated, however, that I had to clean up a bit just to understand what was happening.

While I was in there I also removed the quoting from the 'Overdue Action Type' column.