Bug on release branch OFJ-639 An error occurs when only click the open button on workflow tab
Wrap incident classification in try-catch. Throw Fisma_Exception_User if the category is blank and tell user to pick a category.

Fix the addMultiOption call which does the wrong thing but bafflingly doesn't generate any warnings.

Lastly, add a check for blank comments, since this is a convenient time to fix that, too.