Bug on release branch OFJ-673 XSS Listener Not firing on all columns
Add a quick dirty fuzzing script. This isn't super useful on the branch, since we know that we don't handle bad inputs well, but should be more useful on the trunk once we get it merged back.

It works by overwriting any data which already exists, so empty tables will not be fuzzed.

It also only overwrites the first 25 rows of any table, because there is no need to overwrite an entire table.