Improvement on trunk OFJ-396 Restructure menus
The goal of this is to make each module more... "modular". In this step, I am moving the administration and report items out of the top-level Report and Administration menus and moving them underneat the module they belong to.

In the case of Report, there is nothing left so I have removed it entirely. Everything in ReportController was migrated to a new controller called FindingReportController.

(Plugins are temporarily living under Finding->Reports->Plugin Reports. This is not ideal but it will have to do for now.)

To govern the new menus, I removed two deprecated privileges for 'report area' and 'configuration area'. In place of these, I have created several new area privileges such as 'finding report area' and 'system inventory admin area'.