Improvement on trunk OFJ-655 Single DataTable Implementation
This is a first pass at a data table implementation. It supports local data sources only, not XHR data sources, but is designed with an abstract base class so that an XHR version can be added later.

I renamed the 'tab-view' scripts directory to 'yui' so that we can continue to add other yui view scripts in there.
Add a couple basic tests also.