Improvement on trunk OFJ-709 Generic report class
Add an abstraction for column so that we can track column level attributes, like sortablility and formatters.

This will break for a moment since I have changed the interface between Fisma_Report and the Yui_DataTable wrapper. But I want to check in the data table stuff under another ticket.