Bug on trunk OFJ-686 Stack traces in log files don't have any newlines
Since the log message displayed in the interface is already displayed in a div that preserves whitespace, there is no need to add HTML. I rewrote that part of ErrorController to work exclusively in plaintext, which means that the logging doesn't need any further conversion.

While I had it open, I also went into the log setup in Fisma.php and restructured the format of log messages to be more readable, clean up that code, add comments, and add better handling for special cases (like errors during an unauthenticated session, or errors resulting from local scripts that won't have an IP address).