New feature on trunk: OFJ-410 Display CVE/Bugtraq/Xref/etc. on the finding detail page
Added a panel at the bottom of the current finding info tab which displays CVEs, Bugtraqs, and Xrefs.

Add getUrl methods to the model of each to simplify this process.

Finally, updated the relations between finding and these models because it was slightly messed up and throwing errors when i tried to hydrate result sets.