Bug on trunk OFJ-665 The table background color is missed after the first upload the evidence
Remove unneeded curly brackets in mitigation-strategy.phtml.

Artifacts.phtml is ridiculously convoluted. My first step was just to clean up the indenting and standards, then I was able to move some things around so that it will render correctly. This scripts needs a complete re-do, but that is already on the table for 2.8 or 2.9 so not worth doing now.