OFJ-770 Create a foreign key between incident and system, create report for incidents by bureau
Added the UI on the report details screen to set the related organization. This field is not displayed during the report process, since it would leak system data (names and nicknames) to unauthenticated users.

Add dashboard chart for viewing incidents reported in the last 90 days aggregated to the bureau level. Add report showing incidents reported per bureau on a monthly basis for the last year.