OFJ-471 The msgbar shows up many error messages when click Test Configuration button many times on LDAP configurations page
Clean up the previous work on this ticket. It was not easy to revert that commit because other commits had already been piled on top of this one, and it caused conflicts.

Instead, I just refactored the entire thing, including the consumers of this code. Primarily this meant rewriting WaitingSpinner.js into Fisma/Spinner.js and updated all the dependencies on it.