OFJ-5 Dynamic Approval Workflow
Move Finding*Controllers and Remediation controller into the finding
module. Fix links in scripts, and change render() to partial() in
library classes, so that the default module can be defined for the view
scripts needed by various libraries/helpers.

Some things might still be broken by this move, but it looks like that
it has mostly worked.