OFJ-830 Finding dashboard chart is too small
Upgrade flash version that is being checked for so we can take advantage of new features.

Update the security control deficiency chart to make better use of the screen, including making the chart larger, making the axis label font size smaller, and adding a scrollbar.

Also, added click through so you can click on a control ID to get to a listing of open findings matching that control.