OFJ-620 Re-Design Search Feature
Unrelated to search, but this is killing me when I'm debugging: set zend_session options to NOT throw an exception when there is a php session error (in Debug mode only, production mode still throws exceptions).

This is a handy feature in production but it obscures PHP session errors when in developemnt mode.