OFJ-929 Only one message is able to show up via PriorityMessenger
Modify message() to take a third optional parameter called "clear",
which when passed as TRUE, overwrites whatever is already in the message
box. The default behavior is to append to the message box.

Updated relevant calls to message() to pass the clear parameter so that
messages aren't stacked on top of each other.