OFJ-620 Re-Design Search Feature
Fix search issue with nested set on finding summary screen. If you clicked the link next to a rolled up organization, the search results would show hits for the organizations child nodes but not for the organization itself. (Fixed by changing < to <= and > to >= -- i.e. an off by one error.)