OFJ-930 Create join table for vulnerability statuses
Renamed the status controller to vulnerability-resolution controller in
order to match the model name and to overall be less confusing.  Also
updated the privileges and role fixtures to reflect the change.

Updated the generate-vulnerability script with the new OPEN, FIXED,
WONTFIX status types.

Changed "date closed" to "date fixed" in the vulnerability view.