OFJ-971 Fix all codesniff errors
Fixed errors reported by phpcs. Most of these had to do with escaping echo commands in view scripts. I have
escaped them even though most of them were echoing strings that were already escaped, or contained HTML code
and thus could not be. (used $this->escape(~~~, 'none'); )

Also not updating scripts/bin/backup.php to use Zend_Date as it is not part of the Zend framework. Should
this be added to the cs ignore list?