OFJ-620 Re-Design Search Feature
Revert parts 4 and 5 of previous commit.

The original problem was the usage of Array.map() in Fisma.Search.Panel, so I tried to fix this by modifying Array.prototype, but that turns out to be an even bigger mess.

Far easier to use php.js's array_map function.

Parts 1, 2, and 3 of previous commit are still intact.