OFJ-772 RMF Step 1: Information types selection and FIPS 199 categorization
Fixed fipsAction. Somebody previously confused Organization and System,
pulling the fips data and treating the system id as the organization id.
Added ability to add information types to a system. A filter still needs
to be added to the data table so that already assigned types can't be
assigned again.