OFJ-1034 The error message is not user friendly when delete an incident category which is relate to an incident
Added a pre-delete hook to check if a user is trying to delete an organization/system when it has an incident associated with it.
A doctorine database error will no longer print SQL queries to users (when in production mode). The message shown on OFJ-1034 will now only show in deveopment mode.