OFJ-1074 The new created system can't show on the system list under Solr
Add a small hack to system controller to work around the fact that a system will not be indexed correctly when it is created (because it relies on the organization table having a foreign key to it, but it doesn't have a key until it gets saved the first time).

Also fix a small problem in index listener created by my previous commit.