OFJ-1111 Improve Select Useability
Added a third-party NestedDataTable widget.
Modified NestedDataTable widget and CSS so the tr styles would only apply to a NestedDataTable object (as apposed to any DataTable object).
Added a new SecurityControlTable which extends NestedDataTable.
Added a new set of actions to the SA controller to suppor the new data table.