OFJ-1109 Move bootstrap related functionality out of Fisma.php
Move cacheManager out of Fisma.php, and configure in application.ini
Default is to use APC cache, comments in the configuration on how to
change to a file backed cache.

All references to Fisma::getCacheManager() updated to get the
cachemanager resource from the bootstrap.

Test harness bootstrap modified to work properly.