OFJ-1162 Security Warning Appears in IE
Due to a bug in IE8 (and possibly other versions), when an element in
the DOM has a relative URL set via javascript, the browser's parser
considers the reference to be an HTTP request, rather than an HTTPS
request. This is confusing though, since the actual request that the
browser makes is over SSL.

The fix here is to prepend relative URLs which are later set on page
load with the server URL.

The alternative is to fix the "helpers" so that they aren't actually
setting URLs for images, and the images would be declared in the
application's CSS files. Since this is highly error prone and we're in
code freeze, I've decided to commit this fix instead.