OFJ-1165 CLONE -The number of findings are still shown on the "Current Security control Deficiencies" chart when an user does not have right to read any organization/system
Apparently with doctrine, if the ArrayValues was an empty array in the code of;
DocQueryObj->whereIn(field,Name,ArrayValues)

Then the WHERE ~ IN ~ never actually gets added to the query. This makes sense given the following is not a valid SQL query;
select COUNT(id) FROM finding WHERE id IN ();

but when ArrayValues = FindingTable::getOrganizationIds(), or basically a list of all organization IDs that this user can see, then this is problematic.

This issue (OFJ-1165) has been resolved by placing in checks before every Doctrine query for charts. Though I am curious about other places in Openfisma that uses the ->whereIn, are they all being check correctly?