OFJ-1170 The numbers on top of bars are incorrect on finding and incident charts
The numbers are actually being drawn correctly, however 0s are being printed on top of them as they represent the count for other stacks on stacked bar charts, hence why the screen shots in this Jira ticket show 0s, that dont look exactly right.
0s are supposed to be hidden, they were, and apparently the functionality to hide them got broken.
This issue has now been resolved with this commit.