OFJ-1172 Remove the "Family and Control Number" display
There was a lot of talk on how the issue of the 'Current Security Control Deficiencies' (bottom most) chart could be changed in order to work faster.
If we have the Family bars drill down to 'Family & Controle Number' bars, then users loose the ability to click a family bar to search for all findings associated with that family.
So in the end, and from suggestions from others, I have changed the chart options (Display By) to include family names, so users can see either a Family Summary (the default chart), or just a given/selected family and THAT family's controle number totals.