OFJ-1198 All open finding data show on the "Open Findings By Organization" chart when the user has right to read only some of the organizations/systems
Apparently when you have a Doctrine object, and you do;
  $q->where(A)
  $q->andWhereIn(B)
  $q->where(C)
  $q->andWhere(D)

statements A and B get dropped and only C and D are queried. One must change the second ->where to ->andWhere

  $q->where(A)
  $q->wereIn(B)
  $q->andWhere(C)
  $q->andWhere(D)

will query A, B, C and D. This is the cause of this bug and it has now been fixed.