OFJ-1206 An error occurs when upload an evidence size is more than 10M under production mode
Modify .htaccess max_post_size. If you upload a file that's equal to
max_post_size, the CSRF token will get cut out, and then you'll get an
unknown error in the UI. This problem also occured if you uploaded a
file that was exactly 10M, and the max_post_size was set to 10M as well.
This is really an administration problem, and not an application
problem, but I'll set some sane defaults here. If somebody is trying to
upload a file that is more than 20 times the default 10M limit, then I'm
not going to feel bad that an unknown error pops up.