OFJ-1232 The column of "Percentage" is not sort by number on Documentation Compliance Report page
Modify query so that the percentage column is a numeric, and not made
into a string by CONCAT
Modify the formatter so that a % sign is added to the data table view
Modify the Fisma_Report_Column for percentage so that it's passed as a
numeric, and not a string

This has the side effect of removing % signs in the PDF and XLS reports,
but I don't have time to care. This would take several hours of work,
and the column is already labeled "PERCENTAGE". Sorry that it's missing
a % sign now. On a positive note, you can actually sort the percentages
in the data tables in the browser now.