OFJ-1287 Chart bugs (and hot fixes applied to FDIC)
Fixed a bug where the Open Findings By Organization chart on the Findings Dashboard never loads when setting the Display By options to System
on the FDIC servers.
The reason for this was due to how many systems the FDIC has, and how many queries were done for each system (which leads to a execution timeout).
This issue has been fixed by creating a more efficient query for the "Open Findings By Organization" chart for this specific mode.