OFJ-1150 Add an option to hide retired systems on the org-tree page
Add a parameter of $includeDisposalSystem to getOrganizationTree() function in treeDataAction() and moveNodeAction().
Add a parameter of $includeDisposalSystem to getOrganizationsByPrivilegeQuery() function.
Move showTree() to buildTree() so that it can be listened for checkbox.
Move relative context menu code to buildContextMenu().
Add listener for element disposalSystem to tree.phtml.
Add 'disposal' to yuiNode.labelStyle if a system’s sdlcPhase is disposal in renderTreeNodes() function.
Add 'disposal' style for displaying background color of disposal system as pink.