OFJ-1231 Clean up print media stylesheet
IE is not doing well with printing and canvas and chart from the jqPlot library (it seems it
is not accepting the left statement in styles). It looks ugly, and even worse when showing
data-labels (none of them are placed correctly over the canvas).

Even the examples given on the jqPlot website dont print correctly under IE -
( http://www.jqplot.com/tests/pointLabelTests.php )

This commit hides the graphical charts and forces the screen reader table to be
visible on printed pages.