OFJ-545 Add unit tests to verify Doctrine's parsing behavior for ORDER BY clauses, etc.
Might be a little overkill, but writing tests that include a MySQL
dependency is a major headache with our current unit tests setup, so I
refactored this considerably so that we can have alternate bootstrap
functions for different kinds of tests.

In this commit I have the unit tests working, with scaffolding to add
database tests (tests with a MySQL dependency).

Next steps:
1. Flesh out the database test scaffolding.
2. Write the tests I originally wanted to write for OFJ-545 (you'll see
non-working tests in the current commit.)