OFJ-1349 No records found shows when click a column bar on the "Open Findings By Organization" and "Current Security Control Deficiencies" charts of finding dashboard page
The cause for this bug was the resolution of another recent jira ticket that involved escaping of chart-links.
If a chart link points to the literal string foo/bar/#ColumnLabel# then the "#ColumnLabel#" will be replaced with the label of the column the user clicked.
The # characters however were being escaped, thus the replacement stopped working since the escaping of links were done before the replacement of variable names with values.

Moved this code to fire before the escaping to take care of this bug.