OFJ-545 Add unit tests to verify Doctrine's parsing behavior for ORDER BY clauses, etc.
Fix the standards a bit -- I had an abstract class with the wrong name
and put into the wrong location.

Also remove unneeded Case.php file.

Finally, move phpunit.xml to config.xml so that people don't include it
accidentally. It's only useful for generating coverage reports and it
slows down the test runner *significantly*, so it makes sense not to
include it by default.