OFJ-1319 Aesthetic chart improvements
Filing this under OFJ-1319 since OFJ-1380 is now closed, and this code addition wont change the application behavior (right now).

An exception will now be thrown when a column with X stacks it is added to a stacked-bar chart that is expecting Y stacks, and X !== Y
So if a chart is instance, and told each column is going to have 3 stacks/layers, and Fisma_Chart->addColumn() is called to add a column with only 2 layers, this exception is thrown.
This will prevent such errors leading to OFJ-1319 in the future.