OFJ-792 Finding Summary Enhancements
Transform the term "reponsible organization" to just "organization",
because with ofj-792, the "responsibility" is really going to move to
the POC. The organization is really the subject or target of the audit,
but not necessarily it's responsible party anymore.

I did not change the physical name of the column because that would be a
headache to rewrite all of the constraints.