OFJ-1449 SAM iterative development / OFJ-1473 SA Summary Screen
Adding my initial version of the SA summary screen. This includes a completely refactored
tree table class as it's base class.

(Also enabled YUI's "rethrow exceptions" error for easier debugging. I don't know if it's
new or not but it is MUCH easier to debug with that feature turned on.)