OFJ-1528 ACL error when a user with read/poc tries to a view POC that is also a user
Modified so that viewing/editing a user poc from the poc screen stays within the poc controller.

Doing so caused a new problem where modifying a user record from the poc controller was not reindexing
the poc. I moved the "updatePocIndex" code from UserController into the User model, which makes more
sense. I put it in the User's postSave() method, so now it will trigger any time a user is created or
modified.

This is much clearer than before and actually seems to function correclty. Don't know why I didn't do
it this way the first time around.