OFJ-1529 Add drag-and-drop for POC view
Add drag and drop for POCs.

Had to refactor a few things and that affected the other drag/drop trees.

I also fixed a bug that apparently exists in 2.15 (after changing YUI text node to HTML node near
the end of the 2.15 testing) where tree views are not styled correctly (missing icons, and disposal
systems are not shaded pink).