OFJ-1597 Create new point of contact does not work when use it twice on create finding page
Create some abstractions to make it easier to pipe user messages to different message boxes.

I've done this in a backward compatible way so as to minimize ripple effects of changing an API.

This hasn't fixed the bug yet, but should make it easier to fix the bug in the next commit.