OFJ-1597 Create new point of contact does not work when use it twice on create finding page
I came up with a bit of a hack to fix this problem. IE seems to have an invisible element hovering over the page that
other browsers do not have, so I added some that targets IE7 which moves the dialog offscreen after hiding it, and then
moves it back on screen before showing the panel.

I also messed with the character that's used to represent the control for close boxes. I wasn't satisfied with
what we had there...