OFJ-1662 IRTM Enhancements
Add POC to the model. The POC is assigned automatically if the system has an ISSO.

(If the system has more than one ISSO, then the POC is chosen arbitrarily.)

Also, when a POC is set, he/she receives an e-mail.

Also dropped the "actions taken" field from the model and merged any existing content
into the description field.

Fixed a bug where the review report screen was showing the organization ID number instead
of the organization's name.