OFJ-1720 Log migrations output
Convert all output in the CLI system and the migration base class to use Zend_Log.
Info messages are sent to STDOUT, warnings and errors are sent to STDERR. In debug
mode, STDOUT also prints debug messages.

The migration class augments the logger with a log writer to data/logs/migration.log.
This log file will capture the arguments and output of each invocation of the migration
script, which should be exactly what we originally wanted.