OFJ-1723 Developer bar doesn't work in IE7
Fix an error in APC cache. This wasn't *caused* by the developer bar, but removing
the developer bar had the side effect of turning warnings into fatal errors. This
is mostly a good thing, but we *are* going to have some issues that have gone
unfixed for a long time rear their ugly heads now.