OFJ-1720 Log migrations output
Build fixed.

The original design of this was pretty tacky. I changed Fisma_Cli_Abstract#_log from protected to private and added public getter/setter for dependency injection. I removed the protected _createLog which made no sense.

I updated all of the scripts to use $this->getLog() instead of $this->_log and fixed the tests that were broken by using the setter to inject a mock log object.