OFJ-1661 As an administrator I should be able to define different types of information systems
Fixed this bug:
> An error shows after upload an icon.

The form that uploads the icon did not have a CSRF token. I added the CSRF token, but I also spent
some time looking at the CSRF plugin (with Andrew) and we discovered that it was getting stuck in
an infinite loop: the predispatch in the plugin throws an exception, which dispatches the error
controller, which calls predispatch again, which throws another exception, etc.

To fix this, the CSRF plugin will now only throw an exception if it is *not* being called from
the error controller. I also changed it from trigger_error to Fisma_Zend_Exception_User since
this saves an unnecessary step of converting the errors to exceptions.